+++ elasticsearch/file/scripts/start.sh
@@ -1,0 +1,2 @@
+envsubst < config.yml > config-subst.yml
+java -jar file.jar server config-subst.yml