+++ elasticsearch/file/src/test/java/nl/knaw/huc/service/HeaderLinkUtilTest.java
@@ -1,0 +1,18 @@
+package nl.knaw.huc.service;
+
+import org.junit.jupiter.api.Test;
+
+import java.util.UUID;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+public class HeaderLinkUtilTest {
+
+  @Test
+  public void extractId_shouldExtractUuid() {
+    var expected = UUID.randomUUID();
+    var extracted = HeaderLinkUtil.extractUuid("</rest/files/" + expected + ">; rel=\"original\"");
+    assertThat(extracted).isEqualTo(expected);
+  }
+
+}