+++ textrepo-app/src/main/java/nl/knaw/huc/resources/task/FindResource.java
@@ -5,12 +5,8 @@ import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
 import nl.knaw.huc.helpers.ContentsHelper;
-import nl.knaw.huc.resources.rest.DocumentMetadataResource;
-import nl.knaw.huc.resources.rest.DocumentsResource;
-import nl.knaw.huc.resources.rest.FileMetadataResource;
-import nl.knaw.huc.resources.rest.FileVersionsResource;
-import nl.knaw.huc.resources.rest.FilesResource;
-import nl.knaw.huc.resources.rest.TypesResource;
+import nl.knaw.huc.resources.HeaderLink.Rel;
+import nl.knaw.huc.resources.HeaderLink.Uri;
 import nl.knaw.huc.service.task.TaskBuilderFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -22,14 +18,20 @@ import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
 import javax.ws.rs.QueryParam;
-import javax.ws.rs.core.Link;
 import javax.ws.rs.core.Response;
-import javax.ws.rs.core.UriBuilder;
 import java.util.Map;
 
 import static javax.ws.rs.core.HttpHeaders.ACCEPT_ENCODING;
 import static javax.ws.rs.core.MediaType.APPLICATION_JSON;
 import static javax.ws.rs.core.MediaType.APPLICATION_OCTET_STREAM;
+import static nl.knaw.huc.resources.HeaderLink.Rel.ORIGINAL;
+import static nl.knaw.huc.resources.HeaderLink.Rel.UP;
+import static nl.knaw.huc.resources.HeaderLink.Rel.VERSION_HISTORY;
+import static nl.knaw.huc.resources.HeaderLink.Uri.DOCUMENT;
+import static nl.knaw.huc.resources.HeaderLink.Uri.DOCUMENT_METADATA;
+import static nl.knaw.huc.resources.HeaderLink.Uri.FILE;
+import static nl.knaw.huc.resources.HeaderLink.Uri.FILE_METADATA;
+import static nl.knaw.huc.resources.HeaderLink.Uri.FILE_VERSIONS;
 import static javax.ws.rs.core.UriBuilder.fromResource;
 
 /**
@@ -40,18 +42,6 @@ import static javax.ws.rs.core.UriBuilder.fromResource;
 @Path("/task/find/{externalId}")
 @Api(tags = {"task", "find"})
 public class FindResource {
-
-  public static final String REL_ORIGINAL = "original";
-  public static final String REL_TYPE = Link.TYPE;
-  public static final String REL_UP = "up";
-  public static final String REL_VERSION_HISTORY = "version-history";
-
-  private static final UriBuilder DOCUMENTS = fromResource(DocumentsResource.class).path("/{id}");
-  private static final UriBuilder DOCUMENT_METADATA = fromResource(DocumentMetadataResource.class);
-  private static final UriBuilder FILES = fromResource(FilesResource.class).path("/{id}");
-  private static final UriBuilder FILE_METADATA = fromResource(FileMetadataResource.class);
-  private static final UriBuilder FILE_VERSIONS = fromResource(FileVersionsResource.class);
-  private static final UriBuilder TYPES = fromResource(TypesResource.class).path("/{id}");
 
   private static final Logger log = LoggerFactory.getLogger(FindResource.class);
 
@@ -85,8 +75,8 @@ public class FindResource {
 
     return Response
         .ok(result.getMetadata(), APPLICATION_JSON)
-        .link(DOCUMENT_METADATA.build(docId), REL_ORIGINAL)
-        .link(DOCUMENTS.build(docId), REL_UP)
+        .link(DOCUMENT_METADATA.build(docId), ORIGINAL)
+        .link(DOCUMENT.build(docId), UP)
         .build();
   }
 
@@ -116,9 +106,9 @@ public class FindResource {
 
     return Response
         .ok(result.getMetadata(), APPLICATION_JSON)
-        .link(FILE_METADATA.build(fileId), REL_ORIGINAL)
-        .link(FILES.build(fileId), REL_UP)
-        .link(TYPES.build(typeId), REL_TYPE)
+        .link(FILE_METADATA.build(fileId), ORIGINAL)
+        .link(FILE.build(fileId), UP)
+        .link(Uri.TYPE.build(typeId), Rel.TYPE)
         .build();
   }
 
@@ -146,11 +136,12 @@ public class FindResource {
     final var contents = result.getContents();
     log.debug("Got latest version contents: {}", contents);
 
-    return contentsHelper.asAttachment(contents, acceptEncoding)
-                         .link(FILE_VERSIONS.build(result.getFileId()), REL_VERSION_HISTORY)
-                         .link(FILES.build(result.getFileId()), REL_UP)
-                         .link(TYPES.build(result.getTypeId()), REL_TYPE)
-                         .build();
+    return contentsHelper
+        .asAttachment(contents, acceptEncoding)
+        .link(FILE_VERSIONS.build(result.getFileId()), VERSION_HISTORY)
+        .link(FILE.build(result.getFileId()), UP)
+        .link(Uri.TYPE.build(result.getTypeId()), Rel.TYPE)
+        .build();
   }
 
 }