+++ textrepo-app/src/test/java/nl/knaw/huc/resources/rest/ContentsResourceTest.java
@@ -41,7 +41,6 @@ public class ContentsResourceTest {
 
   public static final ResourceExtension resource = ResourceExtension
       .builder()
-      .addProvider(MultiPartFeature.class)
       .addResource(new ContentsResource(new ContentsService(FILE_STORAGE), CONTENTS_HELPER))
       .build();