+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+File indexer, more info in readme: `./elasticsearch/file`
+Origin Link header to indexer fields request
+
+### Changed
+- Resources now send 'Content-Encoding: gzip' when sending gzip'd contents
+- Database queries using 'count' now use 'long' as their Java counterpart (was: int)
+
 ## [v1.10]
 - [TT-631] When yielding contents, return gzip'd content when 'Accept-Encoding: gzip'
 - [TT-645] Resources now send 'Content-Encoding: gzip' when sending gzip'd contents